potentially fixing lag; fixing inf sprint
infinite sprint has a problem where dying causes it to not fire properly, likely due to playerdata not really existing properly and it screws up trying to fire the remote on a nil character. also, potentially fixing the lag caused by (what i think is) the pieces of det. ESP being similarly named to that of the std. one